// apps/main/src/common/infrastructure/database/types/db.d.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;
// ...
export interface WsEntities {
  address: string | null;
  code: string | null;
  country: Generated<string | null>;
  created_at: Generated<Date>;
  created_by: number | null;
  deleted_at: Date | null;
  entity_tag_id: number | null;
  global_id: Generated<number>;
  id: Generated<number>;
  id_satu_sehat: number | null;
  is_open_vial: Generated<number | null>;
  is_puskesmas: Generated<number>;
  is_relocation: Generated<number>;
  is_vendor: Generated<number | null>;
  lat: string | null;
  lng: string | null;
  name: string | null;
  postal_code: string | null;
  program_id: number | null;
  province_id: string | null;
  regency_id: string | null;
  status: Generated<number | null>;
  sub_district_id: string | null;
  type: Generated<number>;
  updated_at: Generated<Date>;
  updated_by: number | null;
  village_id: string | null;
}